Unlock on Disconnect, Lock on Connect
When a peer disconnects, we now send ourselves a message from them requesting that every file be unlocked. This should be ignored by locks not belonging to them.
While sending newly connected peers the initial files, we also send them a lock message if the file is currently locked for us.